+++ lib/my/sources/my_get_line_tab.c
@@ -7,8 +7,8 @@
 
 #include <stddef.h>
 
-int my_strlen(const char *str);
-int my_strncmp(const char *s1, const char *s2, int n);
+size_t my_strlen(const char *str);
+int my_strncmp(const char *s1, const char *s2, size_t n);
 
 int my_get_line_tab(const char **tab, const char *str)
 {
